split theme.ts into 6 parts
