+++ editor/src/uuiui/styles/theme/base.ts
@@ -1,0 +1,16 @@
+import { createUtopiColor } from '../utopi-color-helpers'
+
+export const base = {
+  blue: createUtopiColor('#007AFF'),
+  purple: createUtopiColor('#574BE2'),
+  red: createUtopiColor('#FD003B'),
+  orange: createUtopiColor('#FA5E00'),
+  neonpink: createUtopiColor('#FF00FF'),
+  neongreen: createUtopiColor('hsl(120, 100%, 37%)'),
+  neonyellow: createUtopiColor('#FCFF42'),
+  almostBlack: createUtopiColor('hsl(0,0%,10%)'),
+  white: createUtopiColor('white'),
+  black: createUtopiColor('black'),
+  darkgray: createUtopiColor('hsl(0, 0%, 50%)'),
+  darkorange: createUtopiColor('#D05300'),
+}
